Use cloud-sql-jdbc-socket-factory to connect to CloudSQL
This takes care of making sure Cloud SQL can be accessed, removing
the need to manually whitelist IPs that can access the database.
Currently the Google application default credentials are used to
authenticate. In the future this needs to be configurable, and this
code needs some love in general.

Note that no connection pooling is currently done.