Remove violations emitted when enums and messages were removed
Removing an enum is mostly benign -- as long as it wasn't used anywhere,
If it was used, then there will be fields where the type has changed. We
will catch any compatibility issues when validating the field type changes.
Nonetheless, removing an enum technically breaks users of generated code.

Similarly, removing a message is bening if it isn't referenced elsewhere.
It, of course, means that people who update their schema will no longer be
able to read messages of that type (which might be stored on disk or in
a database).