Workaround issue where we sometimes allow changes that broke existing protos
Consider the case where the following files exists in the repository:

```
herp/derp.proto:
  package herp;
  message Derp {}
foo/bar.proto
  package foo;
  import "herp/derp.proto"
  message Bar {
    Derp derp = 1;
  }
```

And a publish request that changes `herp/derp.proto' to:

```
package herp {}
message Herpaderp {} // Derp was removed/renamed!
```

That change will break `foo/bar.proto` -- BUT protoc will succeeded if we
only run it with `foo/bar.proto` as the input.

So, we either need to either:
- Run protoc will ALL files as input, or
- Run protoc will ALL files that DEPEND ON a file being changed as input, or
- Track dependencies are ensured that types in use are not removed

This change takes the first approach to solve the issue. Which isn't ideal,
but was the quickest fix.

To make that change this commit includes changes to refactor DescriptorBuilder
to return `FileDescriptorSet` protos instead of `DescriptorSet`.